login and register finished
