--- src/main/java/com/example/serverJustPoteito/user/User.java
@@ -1,125 +1,0 @@
-package com.example.serverJustPoteito.user;
-
-import com.example.serverJustPoteito.role.Role;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "USER")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    @Column(length = 70)
-    private String name;
-    @Column(length = 120)
-    private String surnames;
-    @Column(length = 70)
-    private String userName;
-    @Column(length = 70)
-    private String email;
-    @Column(length = 70)
-    private String password;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "role_id", foreignKey = @ForeignKey(name = "Fk_role_id"))
-    @JsonManagedReference
-    @JsonIgnoreProperties({ "hibernateLazyInitializer", "handler" })
-    private Role role;
-
-    @Column(name = "role_id", insertable = false, updatable = false)
-    private Integer roleId;
-
-    public User() {
-    }
-
-    public User(Integer id, String name, String surnames, String userName, String email, String password, Role role, Integer roleId) {
-        this.id = id;
-        this.name = name;
-        this.surnames = surnames;
-        this.userName = userName;
-        this.email = email;
-        this.password = password;
-        this.role = role;
-        this.roleId = roleId;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSurnames() {
-        return surnames;
-    }
-
-    public void setSurnames(String surnames) {
-        this.surnames = surnames;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public Role getRole() {
-        return role;
-    }
-
-    public void setRole(Role role) {
-        this.role = role;
-    }
-
-    public Integer getRoleId() {
-        return roleId;
-    }
-
-    public void setRoleId(Integer roleId) {
-        this.roleId = roleId;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", surnames='" + surnames + '\'' +
-                ", userName='" + userName + '\'' +
-                ", email='" + email + '\'' +
-                ", password='" + password + '\'' +
-                ", role=" + role +
-                ", roleId=" + roleId +
-                '}';
-    }
-}
